refactor(git): split intent classifier owner
